+++ docs/CONFIG.md
